django-url-shortener-rest-api/pyproject.toml
[tool.poetry]
name = "url-shortner"
version = "0.1.0"
description = "A simple Postgres backed URL shortener made using Django REST framework."
authors = ["Harsh Navdhare <navdhareharsh2001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.4.1"
dj-database-url = "0.5.0"
Django = "3.2.5"
djangorestframework = "3.12.4"
gunicorn = "20.1.0"
Markdown = "3.3.4"
psycopg2-binary = "2.9.1"
pytz = "2021.1"
sqlparse = "0.4.1"
whitenoise = "5.2.0"
dnspython = "2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
